chore: extract MultiSelect
